New minigame queue should work.
The revamped minigame queues allow for games to start with a range of different queue sizes. Queues that have reached the max number of players per game type will automatically start. If the queue has 2 or more player in, the queue will wait until either the queue fills up, or the timer hits 0. This functionality is similar to Hypixel, except that this method will keep you in the main lobby instead of a separate minigame lobby.